==> step3_create_denominators_lact_dim_season.qmd <==
---
title: "Denominators Calculation"
format: html
editor: source
execute:
    echo: false
    message: false
    warning: false
---

```{r}
library(tidyverse)
library(arrow)
library(dtplyr)
library(zoo)

source('functions/DT_base.R')

```

```{r}
#denominator base files -------------------
animals<-read_parquet('data/intermediate_files/animals.parquet') #each row is an animal

animal_lactations<-read_parquet('data/intermediate_files/animal_lactations.parquet') #each row is an animal lactation


#---------------------------
data_pull_min<-min(animals$data_pull_date_min)
data_pull_max<-max(animals$data_pull_date_max)
```


# Gerard's Attempt - by lactation

```{r}
# function for above to allow date range 

# setes data to be use
label_deno <- "2024"
date_start_deno <- "2024-01-01"
date_end_deno <- "2024-12-31"

# data set create function
fxn_create_deno <- function(label = label_deno,
                            date_start = date_start_deno,
                            date_end = date_end_deno){
# needs error checks  
  time_period_text <- label
date_start <- ymd(date_start)
date_end <- ymd(date_end)


# to get cows with dates in window
deno <- animal_lactations |>
  select(id_animal, 
         id_animal_lact, 
         lact_number, lact_group, lact_group_basic,
         date_fresh, date_archive, date_archive_max) |> 
  # get cull date
  group_by(id_animal) |> 
  mutate(date_left = max(date_archive_max)) |> 
  ungroup() |> 
  # create 5 lact groupp
  mutate (lact_group_5 = case_when(
      (lact_number == 0) ~ 'Heifer', 
      (lact_number == 1) ~ 'LACT 1',
      (lact_number == 2) ~ 'LACT 2',
      (lact_number == 3) ~ 'LACT 3',
      (lact_number == 4) ~ 'LACT 4',
      (lact_number >4) ~ 'LACT 5+',
      .default = 'Unknown')
  ) |> 
  #this works for test, but we need to come back to it to see if can integrate into workflow
  mutate(time_period = time_period_text) |> 
  # removes heifers
  filter(lact_number>0) |>  
  # keeps cows with not culled or cows fresh before end
  # filter(is.na(date_left) & date_fresh <= date_end | date_fresh <= date_end) |> 
  filter(date_fresh <= date_end) |>
  filter(is.na(date_archive) | date_archive >=  date_start) |> 
  # filters cows fresh before start date
  filter(date_fresh <= date_start & date_archive >= date_start |
           date_fresh <= date_start & is.na(date_left) |
  # filters cows fresh after start date
    date_fresh >= date_start & date_archive <= date_end |
           date_fresh >= date_start & is.na(date_left)
         )
}

# test example
denominators <- fxn_create_deno()
 
```


```{r}
# denon processing function

fxn_calc_deno <- function(data = denominators){
  
# create lactation groups denominators
deno_all_lacts <- data |> 
  group_by(time_period, lact_group_basic) |>  
  summarize(count_animals = n_distinct(id_animal), 
            count_lactations = n_distinct(id_animal_lact))|>
  ungroup()%>%
 rename(`Lactation Group` = lact_group_basic )

deno_by_lact_group <- data |>
  group_by(time_period,  lact_group)|>
  summarize(count_animals = n_distinct(id_animal), 
            count_lactations = n_distinct(id_animal_lact)) |>
  ungroup() %>%
 rename(`Lactation Group` = lact_group)

# create more lact groups
deno_by_lact_group_5 <- data |>
  group_by(time_period,  lact_group_5)|>
  summarize(count_animals = n_distinct(id_animal), 
            count_lactations = n_distinct(id_animal_lact)) |>
  ungroup() %>%
  rename('Lactation Group' = lact_group_5)

herd_deno <- bind_rows(deno_all_lacts, 
                  deno_by_lact_group, 
                  deno_by_lact_group_5) |> 
  distinct()

}

herd_deno <- fxn_calc_deno(denominators)

# save file
write_parquet(herd_deno, 'data/intermediate_files/herd_denominators.parquet')



```

# Table of Herd denominator data
This is data from the **herd_denominators.parquet** in intermediate files

```{r}
fxn_DT_base(herd_deno)
```


# Example report

```{r}
# can do multiple years or time periods by combining the above

denominators_yr1 <- fxn_create_deno(label = "2023",
                        date_start = "2023-01-01",
                        date_end = "2023-12-31") |> 
  fxn_calc_deno()

denominators_yr2 <- fxn_create_deno(label = "2024",
                        date_start = "2024-01-01",
                        date_end = "2024-12-31") |> 
  fxn_calc_deno()

deno_2yrs <- bind_rows(denominators_yr1, denominators_yr2)

ggplot(data = deno_2yrs, aes(x = `Lactation Group`, y = count_animals,
                             fill = `Lactation Group`))+
  geom_col()+
  scale_fill_viridis_d()+
  facet_wrap(vars(time_period))+
  theme_minimal()+
  theme(legend.position = "",
        axis.text.x = element_text(angle = 45, hjust = 1)
        )+
  labs(x = "")
  


```

# DIM calculations

```{r}
# function to create data set used to calc DIM denominators

interval_days = 30

# 13 categories gives upto 390
fxn_dim_data <- function(data, date_start = date_start_deno,
                                  date_end = date_end_deno,
                         interval = interval_days) {
  # convert dates
  date_start <- ymd(date_start)
  date_end <- ymd(date_end)

  data <- data |> 
    mutate(dim_start = as.numeric(case_when(date_fresh > date_start ~ NA,
                                            date_fresh < date_start ~ 
                                              date_start - date_fresh)
                                  ),
           dim_end = as.numeric(case_when(is.na(date_archive)~ 
                                            date_end - date_fresh,
                                          # for cows culled after
                                          date_archive > date_end ~
                                            date_end - date_fresh,
                                          .default = date_archive - date_fresh)
                                ),
           dim_start = if_else(is.na(dim_start), 0, dim_start),
           dim_start_cat = ceiling(dim_start / interval),
           dim_end_cat = ceiling(dim_end / interval)
           )
}

fxn_dim_groups <- function(data, num_categories = 13, 
                                  interval = interval_days){
  
  # Generate `dim_` categories dynamically
  for (i in seq_len(num_categories)) {
    col_name <- paste0("dim_", (i-1) * interval, "_", i * interval)
    data <- data |> 
      mutate(!!col_name := if_else(dim_start_cat <= i & dim_end_cat >= i, 1, 0))
  }
  
    # Add the n+1 th category for cows with DIM > (num_categories * interval)
  col_name_extra <- paste0("dim_", num_categories * interval, "_+")
  data <- data |> 
    mutate(!!col_name_extra := if_else(dim_end_cat > num_categories, 1, 0))
  
  return(data)
}


# create data set
deno_dim <- fxn_create_deno() |> 
  fxn_dim_data() |> 
  fxn_dim_groups(num_categories = 13)  

# fxn to calculate
fxn_deno_dim <- function(data = deno_dim, lact_group_var){
  
  data <- data |> 
  select(-c(dim_start, dim_end, dim_end_cat, dim_start_cat)) |> 
  group_by(time_period, {{ lact_group_var }}) |> 
  summarise(across(contains("dim_"), ~ sum(. == 1, na.rm = TRUE))) |> 
  ungroup() |> 
  pivot_longer(cols = -c(time_period, {{ lact_group_var }}), 
               names_to = "DIM", 
               values_to = "count") |> 
  rename(`Lactation Group` = {{ lact_group_var }} )
  
}

# create dim df
deno_dim_all <- fxn_deno_dim(deno_dim, lact_group_basic)
deno_dim_3 <- fxn_deno_dim(deno_dim, lact_group)
deno_dim_5 <- fxn_deno_dim(deno_dim, lact_group_5)


herd_deno_dim <- bind_rows(deno_dim_all, 
                  deno_dim_3, 
                  deno_dim_5) |> 
  distinct() |> 
  #set up for later graphing
  mutate(DIM = str_remove(DIM, "dim_"),
         DIM = fct_inorder(DIM)
         )


# save data
write_parquet(herd_deno_dim, 'data/intermediate_files/herd_dim_denominators.parquet')
```
# Table of DIM denominator data
This is data from the **herd_dim_denominators.parquet** in intermediate files

```{r}
fxn_DT_base(herd_deno_dim)
```


# Example report

```{r}

# test graph
ggplot(data = herd_deno_dim)+
  aes(x = DIM, y = count,
      fill = DIM)+
  geom_col()+
  scale_fill_viridis_d()+
  facet_wrap(vars(`Lactation Group`))+
  theme_minimal()+
  theme(legend.position = "",
        axis.text.x = element_text(angle = 45, hjust = 1)
        )+
  labs(x = "",
       y = "")

```




# Season 

```{r}

# to caculate inventory for each month
# this works but long and Sara's is better

date_grouping <- "month"


# developing code
data <- fxn_create_deno()
  
  # convert dates
  date_start <- ymd(date_start_deno)
  date_end <- ymd(date_end_deno)

  data <- data |> 
    # assign units (season or month)
    mutate(unit_start = case_when(date_fresh > date_start ~ 
                                                floor_date(date_fresh,
                                                          unit = date_grouping),
                                            date_fresh <= date_start ~ 
                                              floor_date(date_start, ,
                                                           unit = date_grouping)
                                  ),
           unit_end = case_when(is.na(date_archive) ~ 
                                  floor_date(date_end,
                                             unit = date_grouping),
                                date_archive >= date_end ~
                                floor_date(date_end,
                                           unit = date_grouping),
                                 .default = floor_date(date_archive, ,
                                                          unit = date_grouping)
                                 ),
           # extracts month and year to allow graphing across years
           unit_start_cat = as.yearmon(unit_start),
           unit_end_cat = as.yearmon(unit_end)
             )

data_class <- data |> 
  mutate(unit_1 = if_else(unit_start_cat <= as.yearmon(date_start_deno) &
                                                     unit_end_cat >=
                                                     as.yearmon(date_start_deno),
                                                   1,0),
         unit_2 = if_else(unit_start_cat <= as.yearmon(date_start_deno) + 1/12 &
                            unit_end_cat >=
                            as.yearmon(date_start_deno) + 1/12,
                                                   1,0),
          unit_3 = if_else(unit_start_cat <= as.yearmon(date_start_deno) + 2/12 &
                            unit_end_cat >=
                            as.yearmon(date_start_deno) + 2/12,
                                                   1,0),
         unit_4 = if_else(unit_start_cat <= as.yearmon(date_start_deno) + 3/12 &
                            unit_end_cat >=
                            as.yearmon(date_start_deno) + 3/12,
                                                   1,0),
         unit_5 = if_else(unit_start_cat <= as.yearmon(date_start_deno) + 4/12 &
                            unit_end_cat >=
                            as.yearmon(date_start_deno) + 4/12,
                                                   1,0),
         unit_6 = if_else(unit_start_cat <= as.yearmon(date_start_deno) + 5/12 &
                            unit_end_cat >=
                            as.yearmon(date_start_deno) + 5/12,
                                                   1,0),
         unit_7 = if_else(unit_start_cat <= as.yearmon(date_start_deno) + 6/12 &
                            unit_end_cat >=
                            as.yearmon(date_start_deno) + 6/12,
                                                   1,0),
         unit_8 = if_else(unit_start_cat <= as.yearmon(date_start_deno) + 7/12 &
                            unit_end_cat >=
                            as.yearmon(date_start_deno) + 7/12,
                                                   1,0),
         unit_9 = if_else(unit_start_cat <= as.yearmon(date_start_deno) + 8/12 &
                            unit_end_cat >=
                            as.yearmon(date_start_deno) + 8/12,
                                                   1,0),
         unit_10 = if_else(unit_start_cat <= as.yearmon(date_start_deno) + 9/12 &
                            unit_end_cat >=
                            as.yearmon(date_start_deno) + 9/12,
                                                   1,0),
         unit_11 = if_else(unit_start_cat <= as.yearmon(date_start_deno) + 10/12 &
                            unit_end_cat >=
                            as.yearmon(date_start_deno) + 10/12,
                                                   1,0),
         unit_12 = if_else(unit_start_cat <= as.yearmon(date_start_deno) + 11/12 &
                            unit_end_cat >=
                            as.yearmon(date_start_deno) + 11/12,
                                                   1,0)
         )

data_class_count <- data_class |> 
  select(-c(unit_start, unit_end, unit_start_cat, unit_end_cat)) |> 
  group_by(time_period, lact_group_basic) |> 
  summarise(across(contains("unit_"), ~ sum(. == 1, na.rm = TRUE))) |> 
  ungroup() |> 
  pivot_longer(cols = -c(time_period, lact_group_basic), 
               names_to = "Month", 
               values_to = "count") |> 
  rename(`Lactation Group` = lact_group_basic) 


# to do make code give labels to unit columns so they are the names of vars
# unit_start_deno
# and allow code bo be flexible in case the time period is shorter like season
# or  6 months




```


```{r}
# sara's code

data <- fxn_create_deno()
  
  # convert dates
  date_start <- ymd(date_start_deno)
  date_end <- ymd(date_end_deno)



#                             Monthly Calculations:
# Generate a list of all months in the date range
months_list <- ymd(seq(date_start, date_end, by = "month")) 

# Define a function to calculate monthly active cows by lactation group
count_active_by_month <- function(month, data = data) {
  # Define the start and end date for this month
  # month_start <- ymd(month)
  month_start <- ymd(month)
  month_end <- ceiling_date(month_start, "month") - days(1)
  
  # Filter animals that were active in this month
  active_animals <- data |> 
    filter(
      # Check if the cow was active in this month
      (date_fresh <= month_end & is.na(date_archive))
    |
        (date_fresh <= month_end & date_archive >= month_start) 
    |
        (date_fresh <= month_end & is.na(date_archive))
      )
  
  # Count the cows in each lactation group for this month
  active_counts_all <- active_animals |> 
    group_by(lact_group_basic) |> 
    summarise(cows_active = n(), .groups = "drop") |> 
    spread(key = lact_group_basic, value = cows_active, fill = 0) |> 
    ungroup()
  
    active_counts_lact_group <- active_animals |> 
    group_by(lact_group) |> 
    summarise(cows_active = n(), .groups = "drop") |> 
    spread(key = lact_group, value = cows_active, fill = 0) |> 
    ungroup() |> 
      select('LACT 3+')
    
    active_counts_lact_group_5 <- active_animals |> 
    group_by(lact_group_5) |> 
    summarise(cows_active = n(), .groups = "drop") |> 
    spread(key = lact_group_5, value = cows_active, fill = 0) |> 
    ungroup()
  
    active_counts <- bind_cols(active_counts_all, 
                               active_counts_lact_group_5, 
                               active_counts_lact_group)
    
  # Add the month column
  active_counts$month <- month
  
  return(active_counts)
}

# Apply the function to each month in the date range
monthly_data <- purrr::map_dfr(months_list, 
                               count_active_by_month, 
                               data = data) |> 
  select(month, everything()) |> 
  mutate(month = as.yearmon(ymd(month))
         ) |> 
  # pivot to make easier to use data
  arrange(month) %>% 
  mutate(month = yearmon(month),
         month = as.factor(month)) |> 
  pivot_longer(starts_with("LACT"),
               names_to = "Lactation Group",
               values_to = "count") 

# save data
write_parquet(monthly_data, 'data/intermediate_files/herd_season_denominators.parquet')

# # testing code
# # code below works as gets close to DC #'s
# 
# data <- fxn_create_deno(label = "2025",
#                           date_start = "2025-02-11",
#                           date_end = "2025-02-12")
# 
# date_start <- ymd("2025-02-11")
# date_end <- ymd(date_end_deno)
# 
# month_start <- ymd(date_start)
# Month_end <- ceiling_date(month_start, "month") - days(1)
#   
#   
# active_animals <- data |> 
#     filter(
#       # Check if the cow was active in this month
#       (date_fresh <= month_end & is.na(date_archive))
#     |
#         (date_fresh <= month_end & date_archive >= month_start) 
#     |
#         (date_fresh <= month_end & is.na(date_archive))
#       )
#   
#   # Count the cows in each lactation group for this month
#   active_counts_all <- active_animals |> 
#     group_by(lact_group_basic) |> 
#     summarise(cows_active = n(), .groups = "drop") |> 
#     spread(key = lact_group_basic, value = cows_active, fill = 0) |> 
#     ungroup()
#   
#     active_counts_lact_group <- active_animals |> 
#     group_by(lact_group) |> 
#     summarise(cows_active = n(), .groups = "drop") |> 
#     spread(key = lact_group, value = cows_active, fill = 0) |> 
#     ungroup() |> 
#       select('LACT 3+')
#     
#     active_counts_lact_group_5 <- active_animals |> 
#     group_by(lact_group_5) |> 
#     summarise(cows_active = n(), .groups = "drop") |> 
#     spread(key = lact_group_5, value = cows_active, fill = 0) |> 
#     ungroup()
#   
#     active_counts <- bind_cols(active_counts_all, 
#                                active_counts_lact_group_5, 
#                                active_counts_lact_group)
# 
#     # This ID's cows missing  
# date_class_filter <- data_class |> 
#   filter(unit_1 == 1) |> 
#   anti_join(active_animals)
#     
    

```

# Table of Season denominator data
This is data from the **herd_season_denominators.parquet** in intermediate files

```{r}
fxn_DT_base(monthly_data)
```


# Example Graph

```{r}
# test graph
monthly_data |> 
ggplot()+
  aes(x = month, y = count,
      fill = month)+
  geom_col()+
  scale_fill_viridis_d()+
  facet_wrap(vars(`Lactation Group`))+
  theme_minimal()+
  theme(legend.position = "",
        axis.text.x = element_text(angle = 45, hjust = 1)
        )+
  labs(x = "",
       y = "")
```
